daily_report/report_requirements.txt
pytest
pylint
pandas
psycopg2-binary
python-dotenv